move the unit test to shipment_test.go
